WIP: Check override
